+++ shoop/campaigns/admin_module/forms/__init__.py
@@ -1,0 +1,48 @@
+# This file is part of Shoop.
+#
+# Copyright (c) 2012-2016, Shoop Ltd. All rights reserved.
+#
+# This source code is licensed under the AGPLv3 license found in the
+# LICENSE file in the root directory of this source tree.
+from ._basket import BasketCampaignForm
+from ._basket_conditions import (
+    BasketTotalAmountConditionForm, BasketTotalProductAmountConditionForm,
+    ContactBasketConditionForm, ContactGroupBasketConditionForm,
+    ProductsInBasketConditionForm
+)
+from ._basket_effects import (
+    BasketDiscountAmountForm, BasketDiscountPercentageForm,
+    FreeProductLineForm
+)
+from ._catalog import CatalogCampaignForm
+from ._catalog_conditions import (
+    ContactConditionForm, ContactGroupConditionForm
+)
+from ._catalog_effects import (
+    ProductDiscountAmountForm, ProductDiscountPercentageForm
+)
+from ._catalog_filters import (
+    CategoryFilterForm, ProductFilterForm, ProductTypeFilterForm
+)
+from ._coupon import CouponForm
+
+__all__ = [
+    "BasketCampaignForm",
+    "BasketDiscountAmountForm",
+    "BasketDiscountPercentageForm",
+    "BasketTotalAmountConditionForm",
+    "BasketTotalProductAmountConditionForm",
+    "CatalogCampaignForm",
+    "CategoryFilterForm",
+    "ContactBasketConditionForm",
+    "ContactConditionForm",
+    "ContactGroupBasketConditionForm",
+    "ContactGroupConditionForm",
+    "CouponForm",
+    "FreeProductLineForm",
+    "ProductDiscountAmountForm",
+    "ProductDiscountPercentageForm",
+    "ProductFilterForm",
+    "ProductsInBasketConditionForm",
+    "ProductTypeFilterForm",
+]