+++ shoop/campaigns/admin_module/forms/_catalog_conditions.py
@@ -1,0 +1,21 @@
+# This file is part of Shoop.
+#
+# Copyright (c) 2012-2016, Shoop Ltd. All rights reserved.
+#
+# This source code is licensed under the AGPLv3 license found in the
+# LICENSE file in the root directory of this source tree.
+from shoop.campaigns.models.context_conditions import (
+    ContactCondition, ContactGroupCondition
+)
+
+from ._base import BaseRuleModelForm
+
+
+class ContactGroupConditionForm(BaseRuleModelForm):
+    class Meta(BaseRuleModelForm.Meta):
+        model = ContactGroupCondition
+
+
+class ContactConditionForm(BaseRuleModelForm):
+    class Meta(BaseRuleModelForm.Meta):
+        model = ContactCondition