+++ shoop_tests/admin/test_contact_edit.py
@@ -6,12 +6,13 @@
 # This source code is licensed under the AGPLv3 license found in the
 # LICENSE file in the root directory of this source tree.
 import pytest
+
 from django.contrib.auth import get_user_model
 
-from shoop.admin.forms.widgets import Select2Multiple
+from shoop.admin.forms.fields import Select2MultipleField
 from shoop.admin.modules.contacts.views.edit import ContactBaseForm
 from shoop.core.models import (
-    Contact, CompanyContact, Gender, get_person_contact, PersonContact
+    CompanyContact, Gender, get_person_contact, PersonContact
 )
 from shoop.testing.factories import create_random_company
 from shoop_tests.utils import printable_gibberish
@@ -52,7 +53,7 @@ def test_company_contact_edit_form():
     assert contact_base_form.is_valid(), contact_base_form.errors
     contact = contact_base_form.save()
     assert isinstance(contact, CompanyContact)
-    assert isinstance(contact_base_form.fields["members"].widget, Select2Multiple)
+    assert isinstance(contact_base_form.fields["members"], Select2MultipleField)
 
 
 @pytest.mark.django_db