+++ doc/changelog.rst
@@ -74,6 +74,9 @@ Localization
 Admin
 ~~~~~
 
+- Remove Select2Multiple-widget as unused
+- Use Select2MultipleField for handling contact members
+- Add Select2MultipleField
 - Add account manager for contact
 - Add PersonContact choice widget
 - Add barcode filter to product list view